requirements.txt
streamlit
pandas
matplotlib
groq
sentence-transformers
mysql-connector-python
